+++ cfspopcon/algorithms/__init__.py
@@ -3,13 +3,10 @@ from .core_radiated_power import calc_core_radiated_power
 from .edge_impurity_concentration import calc_edge_impurity_concentration
 from .extrinsic_core_radiator import calc_extrinsic_core_radiator
 from .single_functions import (
-    calc_confinement_transition_threshold_power,
     calc_f_rad_core,
-    calc_greenwald_fraction,
     calc_magnetic_field_on_axis,
     calc_normalised_collisionality,
     calc_peak_pressure,
-    calc_ratio_P_LH,
     calc_rho_star,
     calc_triple_product,
     require_P_rad_less_than_P_in,
