+++ cfspopcon/formulas/metrics/greenwald_density.py
@@ -1,9 +1,11 @@
-"""Operational limits to avoid disruptive regions."""
+"""Calculate the Greenwald density limit."""
 import numpy as np
 
-from ..unit_handling import ureg, wraps_ufunc
+from ...algorithm_class import Algorithm
+from ...unit_handling import ureg, wraps_ufunc
 
 
+@Algorithm.register_algorithm(return_keys=["greenwald_fraction"])
 @wraps_ufunc(
     return_units=dict(greenwald_fraction=ureg.dimensionless),
     input_units=dict(
@@ -24,11 +26,11 @@ def calc_greenwald_fraction(
     Returns:
         :term:`greenwald_fraction` [~]
     """
-    n_Greenwald = calc_greenwald_density_limit.__wrapped__(plasma_current, inverse_aspect_ratio * major_radius)
+    n_Greenwald = calc_greenwald_density_limit.unitless_func(plasma_current, inverse_aspect_ratio * major_radius)
 
     return float(average_electron_density / n_Greenwald)
 
-
+@Algorithm.register_algorithm(return_keys=["greenwald_density_limit"])
 @wraps_ufunc(return_units=dict(nG=ureg.n20), input_units=dict(plasma_current=ureg.MA, minor_radius=ureg.m))
 def calc_greenwald_density_limit(plasma_current: float, minor_radius: float) -> float:
     """Calculate the Greenwald density limit.
@@ -41,21 +43,3 @@ def calc_greenwald_density_limit(plasma_current: float, minor_radius: float) -> 
         nG Greenwald density limit [n20]
     """
     return plasma_current / (np.pi * minor_radius**2)
-
-
-@wraps_ufunc(
-    return_units=dict(troyon_max_beta=ureg.percent),
-    input_units=dict(minor_radius=ureg.m, magnetic_field_on_axis=ureg.T, plasma_current=ureg.MA),
-)
-def calc_troyon_limit(minor_radius: float, magnetic_field_on_axis: float, plasma_current: float) -> float:
-    """Calculate the maximum value for beta, according to the Troyon limit.
-
-    Args:
-        minor_radius: [m] :term:`glossary link<minor_radius>`
-        magnetic_field_on_axis: [T] :term:`glossary link<magnetic_field_on_axis>`
-        plasma_current: [MA] :term:`glossary link<plasma_current>`
-
-    Returns:
-         troyon_max_beta [~]
-    """
-    return 2.8 * plasma_current / (minor_radius * magnetic_field_on_axis)
